backend/requirements.txt
PyQt5
screen_brightness_control
pycaw
librosa
ffmpeg
scipy
requests
pandas
numpy
comtypes
beautifulsoup4
python-dateutil
audiomentations
PyAudio
pyctcdecode
sounddevice
tqdm
transformers
webrtcvad-wheels
yfinance
numpy
future
soundfile
pyctcdecode
https://github.com/kpu/kenlm/archive/master.zip
